extract StrategyPerformanceChart and convert to runes mode
